Initial MacOS Support  * Fixes some issues with notify-send existence not checked.  * Updates link-configs to put configs in the right place and only link configs relevant to the current OS.  * Ensure dangerous ChromeOS specific stuff is removed from .zshrc  * Creates index.macos.sh that is the current entry-point for MacOS  * Update the bootstrap script to support MacOS.  * Update install-fonts.sh to support MacOS  * Update zsh support stuff to support MacOS  * Separate yarn install to shared install script for MacOS and Linux.

+++ configScripts/link-configs.sh
@@ -11,65 +11,81 @@ ln -sf "$(pwd)/configs/.zshrc" ~/.zshrc
 ln -sf "$(pwd)/configs/.p10k.zsh" ~/.p10k.zsh
 ./print.sh "ZSH configs installed!"
 
-# gtkrc-2.0 config
-ln -sf "$(pwd)/configs/.gtkrc-2.0" ~/.gtkrc-2.0
-./print.sh "GTK config installed!"
-
 # Todo Config
 mkdir -p ~/.todo/
 ln -sf "$(pwd)/configs/todo.conf" ~/.todo/config
 ./print.sh "Todo.sh config installed!"
 
-# Systemd Local Units & Overrides
-rm -rf ~/.config/systemd
-ln -sf "$(pwd)/configs/systemd" ~/.config/systemd
-./print.sh "systemd local units & overrides installed!"
-
-# System-wide systemd Config
-sudo rm -f /etc/systemd/journald.conf 
-sudo ln -sf "$(pwd)/configs/etc/systemd/journald.conf" /etc/systemd/journald.conf 
-sudo systemctl restart systemd-journald
-./print.sh "system-wide systemd configs installed!"
+# Sublime Base Folder
+if [[ $OSTYPE == 'darwin'* ]]; then
+    sublimeBaseFolder="$HOME/Library/Application\ Support"
+else
+    sublimeBaseFolder="$HOME/.config"
+fi 
 
 # Sublime Configs
 mkdir -p ~/.config/sublime-text-3/Packages/User
 # shellcheck disable=SC2016
-find "$(pwd)/configs/sublime" -maxdepth 1 -mindepth 1 -print0 | xargs -n1 -0 -I {} bash -c 'ln -sf "{}" "$HOME/.config/sublime-text-3/Packages/User/$(basename "{}")"'
+find "$(pwd)/configs/sublime" -maxdepth 1 -mindepth 1 -print0 | xargs -0 -I {} bash -c "ln -sf \"{}\" \"$sublimeBaseFolder/sublime-text-3/Packages/User/\$(basename \"{}\")\""
 ./print.sh "Sublime configs installed!"
 
 # Sublime Merge Configs
 mkdir -p ~/.config/sublime-merge/Packages/User
 # shellcheck disable=SC2016
-find "$(pwd)/configs/sublime-merge" -maxdepth 1 -mindepth 1 -print0 | xargs -n1 -0 -I {} bash -c 'ln -sf "{}" "$HOME/.config/sublime-merge/Packages/User/$(basename "{}")"'
+find "$(pwd)/configs/sublime-merge" -maxdepth 1 -mindepth 1 -print0 | xargs -0 -I {} bash -c "ln -sf \"{}\" \"$sublimeBaseFolder/sublime-merge/Packages/User/\$(basename \"{}\")\""
 ./print.sh "Sublime merge configs installed!"
 
 # Lazygit Configs
 mkdir -p ~/.config/jesseduffield/lazygit
 ln -sf "$(pwd)/configs/lazygit.config.yml" ~/.config/jesseduffield/lazygit/config.yml
 
-# dnsmasq.d
-sudo rm -rf /etc/dnsmasq.d
-sudo ln -sf "$(pwd)/configs/dnsmasq.d" /etc/dnsmasq.d
-sudo systemctl restart dnsmasq
-./print.sh "dnsmasq.d configs installed!"
-
 # tldr config
 ln -sf "$(pwd)/configs/.tldrrc" ~/.tldrrc
 ./print.sh "tldr config installed!"
 
-INSTALL_ENCRYPTED=$(jq -r '.installThingsWithEncryptedDeps' config.json)
-
-if [[ $INSTALL_ENCRYPTED = 'true' ]]; then
-    ./print.sh "Installing configs relying on encrypted files..."
+if [[ $OSTYPE == 'darwin'* ]]; then
+    ##### MacOS Specific Configs
+    ./print.sh "Installing MacOS Specific Configs"
+else 
+    ##### ChromeOS Specific Configs
+    ./print.sh "Installing ChromeOS Specific Configs"
     
-    # Deluge Configs
-    mkdir -p ~/.config/deluge
-    ln -sf "$(pwd)/configs/deluge/gtk3ui.conf" ~/.config/deluge/gtk3ui.conf
-    ln -sf "$(pwd)/configs/deluge/hostlist.conf" ~/.config/deluge/hostlist.conf
-    ./print.sh "Deluge configs installed!"
+    # gtkrc-2.0 config
+    ln -sf "$(pwd)/configs/.gtkrc-2.0" ~/.gtkrc-2.0
+    ./print.sh "GTK config installed!"
 
-else
-    ./print.sh "Skipping installing configs relying on encrypted files..."
+    # Systemd Local Units & Overrides
+    rm -rf ~/.config/systemd
+    ln -sf "$(pwd)/configs/systemd" ~/.config/systemd
+    ./print.sh "systemd local units & overrides installed!"
+
+    # System-wide systemd Config
+    sudo rm -f /etc/systemd/journald.conf 
+    sudo ln -sf "$(pwd)/configs/etc/systemd/journald.conf" /etc/systemd/journald.conf 
+    sudo systemctl restart systemd-journald
+    ./print.sh "system-wide systemd configs installed!"
+
+    # dnsmasq.d
+    sudo rm -rf /etc/dnsmasq.d
+    sudo ln -sf "$(pwd)/configs/dnsmasq.d" /etc/dnsmasq.d
+    sudo systemctl restart dnsmasq
+    ./print.sh "dnsmasq.d configs installed!"
+
+
+    INSTALL_ENCRYPTED=$(jq -r '.installThingsWithEncryptedDeps' config.json)
+
+    if [[ $INSTALL_ENCRYPTED = 'true' ]]; then
+        ./print.sh "Installing configs relying on encrypted files..."
+        
+        # Deluge Configs
+        mkdir -p ~/.config/deluge
+        ln -sf "$(pwd)/configs/deluge/gtk3ui.conf" ~/.config/deluge/gtk3ui.conf
+        ln -sf "$(pwd)/configs/deluge/hostlist.conf" ~/.config/deluge/hostlist.conf
+        ./print.sh "Deluge configs installed!"
+
+    else
+        ./print.sh "Skipping installing configs relying on encrypted files..."
+    fi
 fi
